refactor: replace map with multimap for HTTP headers
This commit replaces the std::map used for storing HTTP headers with
std::multimap. This change allows for multiple headers with the same
name, which is necessary for certain HTTP headers according to the
HTTP/1.1 specification.

This commit also updates the RequestParser tests to use std::multimap
for headers and adjusts the assertions accordingly.